requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
factory-boy==3.2.1
Faker==13.3.4
psycopg2==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
split-settings==1.0.0
sqlparse==0.4.2
